requirements.txt
Pillow~=10.2.0
torch~=2.2.0
torchvision~=0.17.0
transformers~=4.37.2
einops~=0.7.0
gradio~=4.16.0
timm~=0.9.12
fastapi~=0.109.0